requirements.txt
pandas==2.0.3
scikit-learn==1.3.2
numpy==1.24.4
scipy==1.10.1
matplotlib==3.7.5

torch==1.11.0
torchvision==0.12.0
torchtext==0.12.0
torchaudio==0.11.0
sentence-transformers==2.7.0

skorch==1.0.0
pot==0.9.4
cmake==3.30.2
pykeops==2.2.3
tqdm==4.66.5
attrdict==2.0.1
seaborn==0.13.2
adjustText==1.2.0
h5py==3.11.0
opentsne==1.0.1
watermark==2.4.3
celluloid==0.2.0
munkres==1.1.4

geoopt
git+https://github.com/jeanfeydy/geomloss
git+https://github.com/rossant/ipycache